requirements.txt
# Dependencies necessary for deployment

absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
docopt==0.6.2
GitPython==3.1.18
gitdb==4.0.7
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
jsonpickle==1.5.2
kiwisolver==1.3.1
markdown==3.3.4
matplotlib==3.4.2
munch==2.5.0
numpy==1.22.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
Pillow==8.3.1
protobuf==3.17.3
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacred==0.8.2
six==1.16.0
smmap==4.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1


# EfficientDet pytorch implementation
-e "git+https://github.com/rwightman/efficientdet-pytorch.git@75e16c2f41370b31365a3af0d9bb3d79ad9e1b62#egg=effdet"  # Apache-2.0 License
antlr4-python3-runtime==4.8
omegaconf==2.1.0  # BSD-3-Clause License
PyYAML==5.4.1  # MIT License
timm==0.4.12  # Apache-2.0 License
